Create File_io.py
I've learnt this chapter recently called File I/O and this is my program. I know basic Japanese and I know how to read words in hiragana and katakana. I used functions, conditional expressions and with statement to make this game. 
It returns the highest number you've scored. And it's really amazing.